move benchmarks to own file
